Refactor guardian init commands and add operator provision stub
