+++ crates/hashi-guardian-init/operator-provision.sample.yaml
@@ -1,0 +1,22 @@
+# Sample configuration for the future `operator provision` command.
+#
+# Run with:
+#   cargo run -p hashi-guardian-init -- operator provision --config operator-provision.sample.yaml
+#
+# This command is currently a stub. Eventually it will initialize a fresh
+# withdraw-mode guardian with operator-supplied state before key provisioners
+# submit their shares.
+
+# gRPC endpoint of the withdraw-mode guardian (CEREMONY_MODE=false).
+guardian_endpoint: "http://localhost:3000"
+
+# S3 config for the guardian's log bucket (object-lock enabled).
+guardian_s3:
+  access_key: "AKIA..."
+  secret_key: "..."
+  bucket_info:
+    bucket: "hashi-guardian-logs"
+    region: "us-east-1"
+
+# TODO: add limiter config, committee source, secret-sharing instance,
+# Bitcoin network, MPC master G, build allowlist, and state-hash output fields.
